PHI built for a single polygon at a single observation point;

+++ include/body.h
@@ -1,6 +1,6 @@
 #ifndef BODY_H
 #define BODY_H
 
-MatrixXd body_builder(int z, int c);
+MatrixXd body_builder(float z, int c);
 
 #endif //BODY_H
